requirements.txt
alembic==1.16.2
amqp==5.3.1
annotated-types==0.7.0
async-timeout==5.0.1
better-profanity==0.7.0
billiard==4.2.1
blinker==1.9.0
celery==5.3.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1